+++ src/views/dashboard/metrics/components/metric-menu.vue
