+++ src/views/dashboard/metrics/components/metrics-chart.vue
