+++ src/views/dashboard/metrics/components/metrics-explorer.vue
