+++ src/views/dashboard/metrics/components/prom-ql-editor.vue
