+++ src/views/dashboard/metrics/components/step-selector.vue
